+++ src/radiohead-serial.ts
@@ -1,24 +1,42 @@
 /*
  * NodeJS RadioHead Serial
  *
- * Copyright (C) 2016 Peter Müller <peter@crycode.de> (https://crycode.de/)
+ * Copyright (c) 2017 Peter Müller <peter@crycode.de> (https://crycode.de/)
  *
- * NodeJS module for communication between some RadioHead nodes and NodeJS using
+ * NodeJS module for communication between some RadioHead nodes and Node.js using
  * the RH_Serial driver of the RadioHead library.
  */
-export class RadioHeadSerial {
+
+import {EventEmitter} from 'events';
+import * as Promise from 'bluebird';
+
+export class RadioHeadSerial extends EventEmitter {
 
   /**
-   * The loaded NodeJS addon.
+   * The maximum message length supported by the RH_Serial driver.
+   * This is the largest supported size of a rx or tx buffer.
    */
-  private _addon:RH_Serial.Addon;
+  public static readonly MAX_MESSAGE_LEN:number = 60;
+
+  /**
+   * The loaded Node.js addon.
+   */
+  private _addon:RadioHeadSerial.Addon;
 
   /**
    * Indicator if the worker is active.
    */
   private workerActive:boolean = false;
 
+  /**
+   * Constructor for a new instance of this class.
+   * @param  {string} port    The serial port/device to be used for the communication. (e.g. /dev/ttyUSB0)
+   * @param  {number} baud    The baud rate to be used for the communication. (e.g. 9600)
+   * @param  {number} address The address of this node in the RadioHead network. Address range goes from 1 to 254.
+   */
   constructor(port:string, baud:number, address:number){
+    super();
+
     // load the addon
     this._addon = require('../build/Release/radiohead-serial.node');
 
@@ -26,56 +44,108 @@ export class RadioHeadSerial {
   }
 
   /**
-   * Start the worker for receiving and sending data.
-   * If the worker already active, an error is thrown.
-   *
-   * @param onRecvCallback Callback which is called when a new message is revcived.
+   * Function for receiving a new message from the worker.
+   * Emits a 'data' event with the type of <RadioHeadSerial.ReceivedData>.
+   * @param {Error}  err    An Error or undefined if no error occured.
+   * @param {number} length The length of the received data.
+   * @param {number} from   The from address of the received message.
+   * @param {number} to     The to address of the received message.
+   * @param {number} id     The id of the received message.
+   * @param {number} flags  The flags of the received message.
+   * @param {Buffer} data   The received data as a Buffer.
    */
-  public start(onRecvCallback:(err:Error, from:number, length:number, data:Buffer)=>void):void{
+  private messageReceived(err:Error, length:number, from:number, to:number, id:number, flags:number, data:Buffer):void{
+    this.emit('data', <RadioHeadSerial.ReceivedData>{
+      error: err,
+      length: length,
+      from: from,
+      to: to,
+      id: id,
+      flags: flags,
+      data: data || new Buffer(0)
+    });
+  }
+
+  /**
+   * Start the worker for receiving and sending data.
+   * If the worker already active, nothing is done.
+   * Emits a 'started' event if the worker has been started.
+   */
+  public start():void{
     if(this.workerActive){
-      throw new Error('already started');
+      return;
     }
 
-    this._addon.start(onRecvCallback);
+    this._addon.start(this.messageReceived.bind(this));
     this.workerActive = true;
+
+    this.emit('started');
   }
 
   /**
    * Stop the worker for receiving and sending data.
-   * If the worker is not active, the callback is immediately called.
-   *
-   * @param callback Callback which is called when the worker has been stopped.
+   * If the worker is not active, the Promise will be resolved immediately.
+   * Emits a 'stopped' event if the worker has been stopped.
+   * @return {Promise} A Promise which will be resolved when the worker has been stopped.
    */
-  public stop(callback:()=>void):void{
+  public stop():Promise<{}>{
     if(!this.workerActive){
-      callback();
-      return;
+      return Promise.resolve(undefined);
     }
 
-    this.workerActive = false;
-    this._addon.stop(callback);
+    return new Promise((resolve:()=>void)=>{
+      this.workerActive = false;
+      this._addon.stop(()=>{
+        this.emit('stopped');
+        resolve();
+      });
+    });
+  }
+
+  /**
+   * Returns true if the worker is active.
+   * @return {boolean} Current state of the worker.
+   */
+  public isWorkerActive():boolean{
+    return this.workerActive;
   }
 
   /**
    * Send a message through the RadioHead network.
-   *
-   * @param to Recipient address. Use 255 for broadcast messages.
-   * @param length Number ob bytes to send from the buffer.
-   * @param data Buffer containing the message to send.
-   * @param callback Callback called after the message is send. First Argument is a possible Error object.
+   * If the worker is not active, the promise will be immediately rejected with an error.
+   * @param  {number} to       Recipient address. Use 255 for broadcast messages.
+   * @param  {Buffer} data     Buffer containing the message to send.
+   * @param  {number} length   Optional number ob bytes to send from the buffer. If not given the whole buffer is sent.
+   * @return {Promise}         A Promise which will be resolved when the message has been sent, or rejected in case of an error.
    */
-  public send(to:number, length:number, data:Buffer, callback:(err:Error)=>void):void{
+  public send(to:number, data:Buffer, length?:number):Promise<{}>{
     if(!this.workerActive){
-      throw new Error('Worker not active');
+      return Promise.reject(new Error('Worker not active'))
     }
 
-    this._addon.send(to, length, data, callback);
+    if(!length){
+      length = data.length;
+    }
+
+    if(length <= 0){
+      return Promise.reject(new Error('Nothing to send'))
+    }
+
+    return new Promise((resolve:()=>void, reject:(err:Error)=>void)=>{
+      this._addon.send(to, length, data, (err:Error)=>{
+        if(err){
+          reject(err);
+        }else{
+          resolve();
+        }
+      });
+    });
+
   }
 
   /**
    * Set the address of this node in the RadioHead network.
-   *
-   * @param address The new address.
+   * @param {number} address The new address.
    */
   public setAddress(address:number):void{
     this._addon.setAddress(address);
@@ -85,8 +155,7 @@ export class RadioHeadSerial {
    * Sets the maximum number of retries.
    * Defaults to 3 at construction time.
    * If set to 0, each message will only ever be sent once.
-   *
-   * @param count New number of retries.
+   * @param {number} count New number of retries.
    */
   public setRetries(count:number):void{
     this._addon.setRetries(count);
@@ -95,6 +164,7 @@ export class RadioHeadSerial {
   /**
    * Returns the currently configured maximum retries count.
    * Can be changed with setRetries().
+   * @return {number} The currently configured maximum retries count.
    */
   public getRetries():number{
     return this._addon.getRetries();
@@ -104,8 +174,7 @@ export class RadioHeadSerial {
    * Sets the minimum retransmit timeout in milliseconds.
    * If an ack is taking longer than this time, a message will be retransmitted.
    * Default is 200.
-   *
-   * @param timeout New timeout in milliseconds.
+   * @param {number} timeout New timeout in milliseconds.
    */
   public setTimeout(timeout:number):void{
     this._addon.setTimeout(timeout);
@@ -114,6 +183,7 @@ export class RadioHeadSerial {
   /**
    * Returns the number of retransmissions we have had to send since starting
    * or since the last call to resetRetransmissions().
+   * @return {number} The number of retransmissions we have had to send since starting.
    */
   public getRetransmissions():number{
     return this._addon.getRetransmissions();
@@ -124,5 +194,13 @@ export class RadioHeadSerial {
    */
   public resetRetransmissions():void{
     this._addon.resetRetransmissions();
+  }
+
+  /**
+   * Tells the receiver to accept messages with any to address, not just messages addressed to this node or the broadcast address.
+   * @param {boolean} promiscuous true if you wish to receive messages with any to address. (default false)
+   */
+  public setPromiscuous(promiscuous:boolean):void{
+    this._addon.setPromiscuous(promiscuous);
   }
 }