+++ typings/index.d.ts
@@ -1,2 +1,2 @@
-/// <reference path="globals/node/index.d.ts" />
+/// <reference path="../node_modules/@types/node/index.d.ts" />
 /// <reference path="radiohead-serial.d.ts" />