+++ typings/radiohead-serial.d.ts
@@ -1,12 +1,16 @@
 /*
  * NodeJS RadioHead Serial
  *
- * Copyright (C) 2016 Peter Müller <peter@crycode.de> (https://crycode.de/)
+ * Copyright (c) 2017 Peter Müller <peter@crycode.de> (https://crycode.de/)
  *
- * NodeJS module for communication between some RadioHead nodes and NodeJS using
+ * NodeJS module for communication between some RadioHead nodes and Node.js using
  * the RH_Serial driver of the RadioHead library.
  */
-declare namespace RH_Serial {
+
+/**
+ * Namespace for interfaces and types for RadioHeadSerial.
+ */
+declare namespace RadioHeadSerial {
 
   /**
    * Interface to the Addon.
@@ -14,7 +18,7 @@ declare namespace RH_Serial {
   interface Addon {
     init(port:string, baud:number, address:number):void;
 
-    start(onRecvCallback:(err:Error, from:number, length:number, data:Buffer)=>void):void;
+    start(onRecvCallback:(err:Error, length:number, from:number, to:number, id:number, flags:number, data:Buffer)=>void):void;
 
     stop(callback:()=>void):void;
 
@@ -31,104 +35,174 @@ declare namespace RH_Serial {
     getRetransmissions():number;
 
     resetRetransmissions():void;
+
+    setPromiscuous(promiscuous:boolean):void;
   }
 
   /**
-   * Options for initialisation of the RadioHeadSerial class.
+   * Data received by the worker.
+   * @type {Object}
    */
-  interface Options {
+  type ReceivedData = {
     /**
-     * Hardware port to be used for serial communiation. (e.g. /dev/ttyUSB0)
+     * An error or undefined if no error occurred.
+     * @type {Error}
      */
-    port:string;
+    error: Error;
 
     /**
-     * Baudrate to be used for serial communiation. (e.g. 9600)
+     * The length of the received data.
+     * @type {number}
      */
-    baud:number;
+    length: number;
 
     /**
-     * Address of this node in the RadioHead network. (e.g. 0x01)
+     * The from address of the received message.
+     * @type {number}
      */
-    address:number;
+    from: number;
+
+    /**
+     * The to address of the received message.
+     * @type {number}
+     */
+    to: number;
+
+    /**
+     * The id of the received message.
+     * @type {number}
+     */
+    id: number;
+
+    /**
+     * The flags of the received message.
+     * @type {number}
+     */
+    flags: number;
+
+    /**
+     * The received data as a Buffer.
+     * @type {Buffer}
+     */
+    data: Buffer;
   }
+
+
+}
+
+/**
+ * The module for RadioHeadSerial.
+ */
+declare module 'radiohead-serial' {
+
+  // Import the EventEmitter
+  import {EventEmitter} from 'events';
+
+  // Import Promise from Bluebird
+  import * as Promise from 'bluebird';
 
   /**
    * The RadioHeadSerial Class.
    */
-  class RadioHeadSerial {
+  export class RadioHeadSerial extends EventEmitter {
+
+    /**
+     * The maximum message length supported by the RH_Serial driver.
+     * This is the largest supported size of a rx or tx buffer.
+     */
+    public static readonly MAX_MESSAGE_LEN:number;
+
+    /**
+     * Constructor for a new instance of this class.
+     * @param  {string} port    The serial port/device to be used for the communication. (e.g. /dev/ttyUSB0)
+     * @param  {number} baud    The baud rate to be used for the communication. (e.g. 9600)
+     * @param  {number} address The address of this node in the RadioHead network. Address range goes from 1 to 254.
+     */
     constructor(port:string, baud:number, address:number);
 
     /**
-     * Start the worker for receiving and sending data.
-     * If the worker already active, an error is thrown.
-     *
-     * @param onRecvCallback Callback which is called when a new message is revcived.
+     * Function for receiving a new message from the worker.
+     * Emits a 'data' event with the type of <RadioHeadSerial.ReceivedData>.
+     * @param {Error}  err    An Error or undefined if no error occured.
+     * @param {number} length The length of the received data.
+     * @param {number} from   The from address of the received message.
+     * @param {number} to     The to address of the received message.
+     * @param {number} id     The id of the received message.
+     * @param {number} flags  The flags of the received message.
+     * @param {Buffer} data   The received data as a Buffer or undefined if no data received.
      */
-    start(onRecvCallback:(err:Error, from:number, length:number, data:Buffer)=>void):void;
+    private messageReceived(err:Error, length:number, from:number, to:number, id:number, flags:number, data:Buffer):void;
+
+    /**
+     * Start the worker for receiving and sending data.
+     * If the worker already active, nothing is done.
+     * Emits a 'started' event if the worker has been started.
+     */
+    public start():void;
 
     /**
      * Stop the worker for receiving and sending data.
-     * If the worker is not active, the callback is immediately called.
-     *
-     * @param callback Callback which is called when the worker has been stopped.
+     * If the worker is not active, the Promise will be resolved immediately.
+     * Emits a 'stopped' event if the worker has been stopped.
+     * @return {Promise} A Promise which will be resolved when the worker has been stopped.
      */
-    stop(callback:()=>void);
+    public stop():Promise<{}>;
 
     /**
      * Send a message through the RadioHead network.
-     *
-     * @param to Recipient address. Use 255 for broadcast messages.
-     * @param length Number ob bytes to send from the buffer.
-     * @param data Buffer containing the message to send.
-     * @param callback Callback called after the message is send. First argument is a possible Error object.
+     * If the worker is not active, the promise will be immediately rejected with an error.
+     * @param  {number} to       Recipient address. Use 255 for broadcast messages.
+     * @param  {Buffer} data     Buffer containing the message to send.
+     * @param  {number} length   Optional number ob bytes to send from the buffer. If not given the whole buffer is sent.
+     * @return {Promise}         A Promise which will be resolved when the message has been sent, or rejected in case of an error.
      */
-    send(to:number, length:number, data:Buffer, callback:(err:Error)=>void):void;
+    public send(to:number, data:Buffer, length?:number):Promise<{}>;
 
     /**
      * Set the address of this node in the RadioHead network.
-     *
-     * @param address The new address.
+     * @param {number} address The new address.
      */
-    setAddress(address:number):void;
+    public setAddress(address:number):void;
 
     /**
      * Sets the maximum number of retries.
      * Defaults to 3 at construction time.
      * If set to 0, each message will only ever be sent once.
-     *
-     * @param count New number of retries.
+     * @param {number} count New number of retries.
      */
-    setRetries(count:number):void;
+    public setRetries(count:number):void;
 
     /**
      * Returns the currently configured maximum retries count.
      * Can be changed with setRetries().
+     * @return {number} The currently configured maximum retries count.
      */
-    getRetries():number;
+    public getRetries():number;
 
     /**
      * Sets the minimum retransmit timeout in milliseconds.
-     * If an ack is taking longer than this time, a message will be retransmittet.
+     * If an ack is taking longer than this time, a message will be retransmitted.
      * Default is 200.
-     *
-     * @param timeout New timeout in milliseconds.
+     * @param {number} timeout New timeout in milliseconds.
      */
-    setTimeout(timeout:number):void;
+    public setTimeout(timeout:number):void;
 
     /**
      * Returns the number of retransmissions we have had to send since starting
      * or since the last call to resetRetransmissions().
+     * @return {number} The number of retransmissions we have had to send since starting.
      */
-    getRetransmissions():number
+    public getRetransmissions():number;
 
     /**
      * Resets the count of the number of retransmissions to 0.
      */
-    resetRetransmissions():void;
-  }
-}
+    public resetRetransmissions():void;
 
-declare module 'radiohead-serial' {
-  export import RadioHeadSerial = RH_Serial.RadioHeadSerial;
+    /**
+     * Tells the receiver to accept messages with any to address, not just messages addressed to this node or the broadcast address.
+     * @param {boolean} promiscuous true if you wish to receive messages with any to address. (default false)
+     */
+    public setPromiscuous(promiscuous:boolean):void;
+  }
 }