+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v2.0.0 DEV
+- New APIv2 with breaking changes to use Events and Promises
+- Added more information for received messages (length, from, to, id, flags)
+- Added support for the promiscuous mode of the receiver, to receive messages with any to address
+- Added constant MAX_MESSAGE_LEN
+- Updated TypeScript typings
+- Added TypeScript examples
+
 ## v1.0.4 2017-03-21
 - Updated RadioHead library to v1.74